move ecs auth/ssl to setup
