+++ docs/ecs-commands.md
@@ -4,7 +4,7 @@ For information about the ECS Server functionality, see the [ecs-server](ecs-ser
 
 ## Commands
 
-### ecs auth
+### setup ecs auth
 
 Configures the HTTP Authentication BearerToken.  Once set, all future client
 requests to the ECS Server will need to provide the correct credentials.  
@@ -17,6 +17,23 @@ Flags:
 
  * `--bearer-token` -- Specify the bearer token secret.
  * `--delete` -- Delete the bearer token and disable authentication.
+
+---
+
+### setup ecs ssl
+
+ Configures the SSL Certificate and Private Key to enable SSL/TLS.  Saves the
+ SSL certificate and private key to the SecureStore.
+
+ **Note:** At this time, this feature is not recommended due to a bug
+ in the [AWS SDK](https://github.com/boto/boto3/issues/4188).
+
+ Flags:
+
+  * `--delete` -- Disables SSL and deletes both the SSL certificate and private key from the Secure Store
+  * `--print` -- Prints the SSL certificate
+  * `--certificate` -- Path to SSL certificate file in PEM format
+  * `--private-key` -- Path to SSL private key in PEM format
 
 ---
 
@@ -76,23 +93,6 @@ Fetches the ProfileName of the role stored in the default slot of the ECS Server
 Flags:
 
  * `--slotted` -- Load the IAM credentials into a unique slot using the ProfileName as the key
-
----
-
-### ecs ssl
-
- Configures the SSL Certificate and Private Key to enable SSL/TLS.  Saves the
- SSL certificate and private key to the SecureStore.
-
- **Note:** At this time, this feature is not recommended due to a bug
- in the [AWS SDK](https://github.com/boto/boto3/issues/4188).
-
- Flags:
-
-  * `--delete` -- Disables SSL and deletes both the SSL certificate and private key from the Secure Store
-  * `--print` -- Prints the SSL certificate
-  * `--certificate` -- Path to SSL certificate file in PEM format
-  * `--private-key` -- Path to SSL private key in PEM format
 
 ---
